warning fixes on windows

+++ pynari/Array.cpp
@@ -29,7 +29,7 @@ namespace pynari {
     py::array_t<T> asArray = py::cast<py::array_t<T>>(buffer);
     int numScalarsInArray = 1;
     for (int i=0;i<info.ndim;i++)
-      numScalarsInArray *= info.shape[i];
+      numScalarsInArray *= (int)info.shape[i];
 
     anari::Array handle = 0;
     if (D == 1) {
